refactor: implement the requested changes
