+++ public/MenuItems/siteconfig.json
@@ -1,0 +1,8 @@
+{
+    "menuItems": [
+      { "name": "Home", "path": "/" },
+      { "name": "About", "path": "/about" },
+      { "name": "Services", "path": "/services" },
+      { "name": "Contact", "path": "/contact" }
+    ]
+  }
